# backend/docs/swagger.yaml
basePath: /
definitions:
...
  models.CreateRoomRequest:
    description: Room creation request data
    properties:
      capacity:
        example: 200
        minimum: 1
        type: integer
      description:
        example: Spacious ballroom with high ceilings
        maxLength: 1000
        type: string
      floor:
        example: 1
        minimum: 0
        type: integer
      name:
        example: Main Ballroom
        maxLength: 255
        minLength: 1
        type: string
      room_type:
        allOf:
        - $ref: '#/definitions/models.RoomType'
        enum:
        - general
        - ballroom
        - conference
        - theater
        - banquet
        - outdoor
        example: ballroom
    required:
    - capacity
    - name
    - room_type
    type: object
...
  models.Event:
    description: Event information and details
    properties:
      allow_self_rsvp:
        description: Guest Management Options
        type: boolean
      categories:
        description: Discovery and Organization (for public events)
        items:
          type: string
        type: array
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      date:
        example: "2024-06-15T18:00:00Z"
        type: string
      description:
        example: A beautiful celebration of love
        type: string
      guests:
        description: Guest Management
        items:
          $ref: '#/definitions/models.Guest'
        type: array
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      location:
        example: Grand Hotel Ballroom
        type: string
      max_guests:
        type: integer
      name:
        example: John & Jane's Wedding
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      require_approval:
        type: boolean
      slug:
        example: john-jane-wedding-2024
        type: string
      tags:
        items:
          type: string
        type: array
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      visibility:
        allOf:
        - $ref: '#/definitions/models.EventVisibility'
        description: Visibility and Access Control
    type: object
...
  models.EventVisibility:
    enum:
    - private
    - public
    type: string
    x-enum-varnames:
    - EventVisibilityPrivate
    - EventVisibilityPublic
...
  models.Guest:
    description: Guest information and RSVP status
    properties:
      approved:
        type: boolean
      created_at:
        description: Timestamps
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      event:
        $ref: '#/definitions/models.Event'
      event_id:
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        description: Guest Information
        example: John Doe
        type: string
      notes:
        example: Vegetarian, allergic to nuts
        type: string
      phone:
        example: "+1234567890"
        type: string
      rsvp_date:
        example: "2024-01-15T10:30:00Z"
        type: string
      rsvp_status:
        allOf:
        - $ref: '#/definitions/models.RSVPStatus'
        description: RSVP Status
      seat:
        $ref: '#/definitions/models.Seat'
      seat_id:
        description: Seating (for future implementation)
        type: string
      source:
        allOf:
        - $ref: '#/definitions/models.GuestSource'
        description: Source and Approval
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      user:
        $ref: '#/definitions/models.User'
      user_id:
        description: User relationship (only if guest is a platform user) - NEW FIELD
        type: string
    type: object
...
  models.GuestSource:
    enum:
    - owner_added
    - invitation
    - user_registration
    type: string
    x-enum-comments:
      GuestSourceInvitation: Registered via invitation
      GuestSourceOwnerAdded: Added by owner manually
      GuestSourceUserRegistration: Self-registered by authenticated platform user
    x-enum-descriptions:
    - Added by owner manually
    - Registered via invitation
    - Self-registered by authenticated platform user
    x-enum-varnames:
    - GuestSourceOwnerAdded
    - GuestSourceInvitation
    - GuestSourceUserRegistration
...
  models.RSVPStatus:
    enum:
    - pending
    - accept
    - decline
    - maybe
    type: string
    x-enum-varnames:
    - RSVPStatusPending
    - RSVPStatusAccept
    - RSVPStatusDecline
    - RSVPStatusMaybe
  models.Room:
    description: Room information and details
    properties:
      capacity:
        example: 200
        type: integer
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Spacious ballroom with high ceilings
        type: string
      floor:
        example: 1
        type: integer
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: Main Ballroom
        type: string
      room_type:
        $ref: '#/definitions/models.RoomType'
      seats:
        items:
          $ref: '#/definitions/models.Seat'
        type: array
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      venue:
        $ref: '#/definitions/models.Venue'
      venue_id:
        type: string
    type: object
  models.RoomResponse:
    description: Room response data
    properties:
      room:
        $ref: '#/definitions/models.Room'
    type: object
  models.RoomType:
    enum:
    - general
    - ballroom
    - conference
    - theater
    - banquet
    - outdoor
    type: string
    x-enum-varnames:
    - RoomTypeGeneral
    - RoomTypeBallroom
    - RoomTypeConference
    - RoomTypeTheater
    - RoomTypeBanquet
    - RoomTypeOutdoor
  models.RoomsResponse:
    description: Multiple rooms response data
    properties:
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      total:
        type: integer
    type: object
  models.Seat:
    description: Seat information and details
    properties:
      category:
        $ref: '#/definitions/models.SeatCategory'
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      guest:
        $ref: '#/definitions/models.Guest'
      guest_id:
        type: string
      height:
        example: 1
        type: number
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      number:
        example: "1"
        type: string
      room:
        $ref: '#/definitions/models.Room'
      room_id:
        type: string
      rotation:
        example: 0
        type: number
      row:
        example: A
        type: string
      status:
        $ref: '#/definitions/models.SeatStatus'
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      width:
        example: 1
        type: number
      x:
        description: Position X coordinate
        example: 10.5
        type: number
      "y":
        description: Position Y coordinate
        example: 15.2
        type: number
    type: object
  models.SeatCategory:
    enum:
    - standard
    - vip
    - accessible
    - premium
    - economy
    - standing
    type: string
    x-enum-varnames:
    - SeatCategoryStandard
    - SeatCategoryVIP
    - SeatCategoryAccessible
    - SeatCategoryPremium
    - SeatCategoryEconomy
    - SeatCategoryStanding
...
  models.SeatStatus:
    enum:
    - available
    - occupied
    - reserved
    - blocked
    - maintenance
    type: string
    x-enum-varnames:
    - SeatStatusAvailable
    - SeatStatusOccupied
    - SeatStatusReserved
    - SeatStatusBlocked
    - SeatStatusMaintenance
...
  models.UpdateRoomRequest:
    description: Room update request data
    properties:
      capacity:
        example: 250
        minimum: 1
        type: integer
      description:
        example: Spacious ballroom with high ceilings
        maxLength: 1000
        type: string
      floor:
        example: 1
        minimum: 0
        type: integer
      name:
        example: Main Ballroom
        maxLength: 255
        minLength: 1
        type: string
      room_type:
        allOf:
        - $ref: '#/definitions/models.RoomType'
        enum:
        - general
        - ballroom
        - conference
        - theater
        - banquet
        - outdoor
        example: ballroom
    type: object
...
  models.User:
    description: User account information
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      first_name:
        example: John
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      last_name:
        example: Doe
        type: string
      phone:
        example: "+1234567890"
        type: string
      profile_pic:
        example: https://example.com/profile.jpg
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
...
  models.Venue:
    description: Venue information and details
    properties:
      address:
        example: 123 Main Street
        type: string
      city:
        example: New York
        type: string
      country:
        example: USA
        type: string
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      description:
        example: Elegant ballroom with modern amenities
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      is_public:
        type: boolean
      name:
        example: Grand Hotel Ballroom
        type: string
      owner:
        $ref: '#/definitions/models.User'
      owner_id:
        type: string
      phone:
        example: +1-555-123-4567
        type: string
      postal_code:
        example: "10001"
        type: string
      rooms:
        items:
          $ref: '#/definitions/models.Room'
        type: array
      state:
        example: NY
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      website:
        example: https://grandhotel.com
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the SeatMaster backend server for event management and seating
    arrangements.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SeatMaster Backend API
  version: "1.0"
paths:
...
  /venues/{id}/rooms:
    get:
      description: Retrieve all rooms in a specific venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomsResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get rooms in venue
      tags:
      - rooms
    post:
      consumes:
      - application/json
      description: Create a new room in a venue
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room creation data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/models.CreateRoomRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.RoomResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create a new room
      tags:
      - rooms
  /venues/{id}/rooms/{roomId}:
    delete:
      description: Delete an existing room
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete room
      tags:
      - rooms
    get:
      description: Retrieve a specific room by ID
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get room by ID
      tags:
      - rooms
    patch:
      consumes:
      - application/json
      description: Update an existing room
      parameters:
      - description: Venue ID
        in: path
        name: id
        required: true
        type: string
      - description: Room ID
        in: path
        name: roomId
        required: true
        type: string
      - description: Room update data
        in: body
        name: room
        required: true
        schema:
          $ref: '#/definitions/models.UpdateRoomRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RoomResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update room
      tags:
      - rooms
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"